@@ -6,7 +6,7 @@
 
 package net.mediaarea.mediainfo
 
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.net.Uri
 import android.content.Intent
