@@ -9,9 +9,9 @@ package net.mediaarea.mediainfo
 import java.io.OutputStream
 import java.io.File
 
-import android.support.v4.app.Fragment
-import android.support.design.widget.Snackbar
-import android.support.v4.provider.DocumentFile
+import androidx.fragment.app.Fragment
+import androidx.documentfile.provider.DocumentFile
+
 import android.os.Build
 import android.os.Bundle
 import android.os.Environment
@@ -21,6 +21,8 @@ import android.content.Context
 import android.content.Intent
 import android.text.TextUtils
 import android.view.*
+
+import com.google.android.material.snackbar.Snackbar
 
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.disposables.CompositeDisposable
