@@ -6,7 +6,7 @@
 
 package net.mediaarea.mediainfo
 
-import android.arch.lifecycle.ViewModel
+import androidx.lifecycle.ViewModel
 import io.reactivex.Completable
 import io.reactivex.Flowable
 import io.reactivex.Single
