@@ -6,9 +6,9 @@
 
 package net.mediaarea.mediainfo
 
-import android.arch.persistence.room.RoomDatabase
-import android.arch.persistence.room.Database
-import android.arch.persistence.room.Room
+import androidx.room.RoomDatabase
+import androidx.room.Database
+import androidx.room.Room
 import android.content.Context
 
 @Database(entities = [Report::class], version = 1, exportSchema = false)
